Update patch set 3
Patch Set 3: Code-Review+1

Thanks for cleaning this up. 

The app name is only displayed in the 'share' picker when you generate a bug report and we wanted to make sure it's obviously the CM JIRA not just a random one. 

The error messages were short because there isn't much room in a standard notification.

Patch-set: 3