Update patch set 1
Patch Set 1: -Code-Review

I'll start making those changes momentarily. 

I was thinking as well that the zipfile.delete() should also be in a finally block but didn't see an easy way to do it unless I want to move the try/catch for io and zip exceptions inside attachFile(). Would this be the best way to do this? It is using the app's cache dir so they'll get purged eventually but even zipped the files are ~1MB in size so I'd rather clean up as it goes.

Patch-set: 1