Update patch set 2
Patch Set 2:

Is it possible to get a different notification icon?

CMBugReport's icon is the same as the icon for USB Debugging, which is a little weird when 2 of them are next to eachother in the statusbar.

Patch-set: 2
Reviewer: Eddy Witkamp <2881@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0