Update patch set 1
Patch Set 1:

I decided on a whim to implement this today. For the apps that I am developing, I see the CM Bug Report come up any time there is an unexpected crash while I'm out and about using my phone.

I don't want to submit these in-development bugs to Jira, but the stack trace is very useful to me, so I want to export it somehow. This way I can just copy paste it wherever I need it to make a note. The feature will be hidden for now, so developers can use it, but it won't get in the way of normal users.

Patch-set: 1